@@ -5,34 +5,34 @@
   "description": "TableTop RPG tools",
   "main": "index.js",
   "scripts": {
+    "lint": "eslint src/",
+    "lint:fix": "eslint --fix src/",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
     "eject": "react-scripts eject",
     "deploy": "npm run build && firebase deploy",
     "scss": "node-sass src/_assets/scss/main.scss src/_assets/css/main.css",
     "scss:watch": "npm run scss && node-sass --watch src/_assets/scss/main.scss --output-style compressed --source-map true --output src/_assets/css",
-    "cover": "npm test -- --coverage"
+    "test:cover": "npm test -- --coverage --watchAll=false --onlyChanged=false",
+    "test:watch-cover": "npm test -- --coverage --onlyChanged=false",
+    "test": "react-scripts test"
   },
   "author": "Joe Ziemba",
   "license": "ISC",
   "dependencies": {
-    "firebase": "^5.11.1",
+    "firebase": "^8.0.0",
     "gh-pages": "^2.2.0",
-    "lodash": "^4.17.15",
-    "react": "^16.12.0",
-    "react-bootstrap-typeahead": "^3.4.7",
-    "react-dom": "^16.12.0",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
-    "react-toastify": "^5.5.0",
-    "reactstrap": "^7.1.0",
-    "sass": "^1.25.0",
-    "typescript": "^3.7.5"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "lodash": "^4.17.21",
+    "react": "^17.0.2",
+    "react-bootstrap-typeahead": "^5.0.0",
+    "react-dom": "^17.0.2",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^3.4.0",
+    "react-toastify": "^7.0.0",
+    "reactstrap": "^8.9.0",
+    "sass": "^1.32.12",
+    "typescript": "^3.9.9"
   },
   "browserslist": [
     ">0.2%",
@@ -41,8 +41,15 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "node-sass": "^4.13.1"
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/react": "^11.2.6",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-jest": "^24.3.6",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "node-sass": "^4.14.1",
+    "prettier": "^2.2.1"
   }
 }
