@@ -21,12 +21,12 @@ class App extends Component {
     super(props);
 
     this.state = {
-      currentUser: null
+      currentUser: null,
     };
   }
 
   componentDidMount() {
-    firebase.auth.onAuthStateChanged(currentUser => {
+    firebase.auth.onAuthStateChanged((currentUser) => {
       currentUser
         ? this.setState({ currentUser })
         : this.setState({ currentUser: null });
@@ -35,7 +35,7 @@ class App extends Component {
 
   render() {
     return (
-      <div className="">
+      <>
         <ToastContainer
           toastClassName="c-toast"
           autoClose={3000}
@@ -45,7 +45,7 @@ class App extends Component {
         <UserContext.Provider
           value={{
             currentUser: this.state.currentUser,
-            registerCurrentUserToState: this.registerCurrentUserToState
+            registerCurrentUserToState: this.registerCurrentUserToState,
           }}
         >
           <Route exact path="/" component={Home} />
@@ -71,7 +71,7 @@ class App extends Component {
             component={CharacterBuilder}
           />
         </UserContext.Provider>
-      </div>
+      </>
     );
   }
 }
