@@ -1,7 +1,7 @@
 import React from "react";
 import Statbox from "./Statbox";
 
-const SelectStatbox = props => {
+const SelectStatbox = (props) => {
   const levels = [];
   for (let i = 1; i < 21; i++) {
     levels.push(i);
@@ -9,8 +9,12 @@ const SelectStatbox = props => {
   return (
     <Statbox title="Level" large>
       <select onChange={props.setLevel} className="c-statbox__select">
-        {levels.map(level => {
-          return <option value={level}>{level}</option>;
+        {levels.map((level, i) => {
+          return (
+            <option value={level} key={i}>
+              {level}
+            </option>
+          );
         })}
       </select>
     </Statbox>
