@@ -10,29 +10,29 @@ export const Ancestries = {
       {
         ability: Abilities.CON,
         source: "Dwarf",
-        id: "1"
+        id: "1",
       },
       {
         ability: Abilities.WIS,
         source: "Dwarf",
-        id: "2"
+        id: "2",
       },
       {
         ability: Abilities.FREE,
         source: "Dwarf",
         id: "3",
-        type: Abilities.FREE
-      }
+        type: Abilities.FREE,
+      },
     ],
     abilityFlaws: [
       {
         ability: Abilities.CHA,
         source: "Dwarf",
-        id: "1"
-      }
+        id: "1",
+      },
     ],
     languages: ["Common", "Dwarven"],
-    traits: ["Dwarf", "Humanoid"]
+    traits: ["Dwarf", "Humanoid"],
   },
   Elf: {
     name: "Elf",
@@ -43,29 +43,29 @@ export const Ancestries = {
       {
         ability: Abilities.DEX,
         source: "Elf",
-        id: "1"
+        id: "1",
       },
       {
         ability: Abilities.INT,
         source: "Elf",
-        id: "2"
+        id: "2",
       },
       {
         ability: Abilities.FREE,
         source: "Elf",
         id: "3",
-        type: Abilities.FREE
-      }
+        type: Abilities.FREE,
+      },
     ],
     abilityFlaws: [
       {
         ability: Abilities.CON,
         source: "Elf",
-        id: "1"
-      }
+        id: "1",
+      },
     ],
     languages: ["Common", "Elven"],
-    traits: ["Elf", "Humanoid"]
+    traits: ["Elf", "Humanoid"],
   },
   Gnome: {
     name: "Gnome",
@@ -79,21 +79,23 @@ export const Ancestries = {
         ability: Abilities.FREE,
         source: "Gnome",
         id: "Gnome3",
-        type: Abilities.FREE
-      }
+        type: Abilities.FREE,
+      },
     ],
-    abilityFlaws: [{ ability: Abilities.STR, source: "Gnome", id: "Gnome1" }],
+    abilityFlaws: [
+      { ability: Abilities.STR, source: "Gnome", id: "Gnome1" },
+    ],
     languages: ["Common"],
     traits: ["Gnome", "Humanoid"],
     heritages: {
       halfElf: {
         name: "Half Elf",
-        traits: ["Elf"]
+        traits: ["Elf"],
       },
       halfOrc: {},
       skilled: {},
-      versatile: {}
-    }
+      versatile: {},
+    },
   },
   Goblin: {
     name: "Goblin",
@@ -104,28 +106,28 @@ export const Ancestries = {
       {
         ability: Abilities.DEX,
         source: "Goblin",
-        id: "1"
+        id: "1",
       },
       {
         ability: Abilities.CHA,
         source: "Goblin",
-        id: "2"
+        id: "2",
       },
       {
         ability: Abilities.FREE,
         source: "Goblin",
         id: "3",
-        type: Abilities.FREE
-      }
+        type: Abilities.FREE,
+      },
     ],
     abilityFlaws: [
       {
         ability: Abilities.WIS,
         source: "Goblin",
-        id: "1"
-      }
+        id: "1",
+      },
     ],
-    languages: ["Common", "Goblin"]
+    languages: ["Common", "Goblin"],
   },
   Human: {
     name: "Human",
@@ -137,14 +139,14 @@ export const Ancestries = {
         ability: Abilities.FREE,
         source: "Human",
         id: "Human1",
-        type: Abilities.FREE
+        type: Abilities.FREE,
       },
       {
         ability: Abilities.FREE,
         source: "Human",
         id: "Human2",
-        type: Abilities.FREE
-      }
+        type: Abilities.FREE,
+      },
     ],
     abilityFlaws: [],
     languages: ["Common"],
@@ -152,12 +154,12 @@ export const Ancestries = {
     heritages: {
       halfElf: {
         name: "Half Elf",
-        traits: ["Elf"]
+        traits: ["Elf"],
       },
       halfOrc: {},
       skilled: {},
-      versatile: {}
-    }
+      versatile: {},
+    },
   },
   Halfling: {
     name: "Halfling",
@@ -168,27 +170,27 @@ export const Ancestries = {
       {
         ability: Abilities.DEX,
         source: "Halfling",
-        id: "1"
+        id: "1",
       },
       {
         ability: Abilities.WIS,
         source: "Halfling",
-        id: "2"
+        id: "2",
       },
       {
         ability: Abilities.FREE,
         source: "Halfling",
         id: "3",
-        type: Abilities.FREE
-      }
+        type: Abilities.FREE,
+      },
     ],
     abilityFlaws: [
       {
         ability: Abilities.STR,
         source: "Halfling",
-        id: "1"
-      }
+        id: "1",
+      },
     ],
-    languages: ["Common", "Halfling"]
-  }
+    languages: ["Common", "Halfling"],
+  },
 };
