@@ -1,6 +1,6 @@
 import React from "react";
 
-const Column = props => {
+const Column = (props) => {
   return <div className={`col ${props.className}`}>{props.children}</div>;
 };
 
