@@ -1,17 +1,17 @@
-import React from 'react';
-import Row from './Row';
-import Column from './Column';
+import React from "react";
+import Row from "./Row";
+import Column from "./Column";
 
 const InputFlat = (props) => {
   return (
     <div className="form-group">
       <Row>
-        <Column className='col'>
+        <Column className="col">
           <label htmlFor={props.id}>{props.label}</label>
         </Column>
-        <Column className='col'>
+        <Column className="col">
           <input
-            type={props.type || 'text'}
+            type={props.type || "text"}
             className="form-control form-control-sm"
             id={props.id}
             aria-describedby={`${props.id}Help`}
@@ -20,12 +20,13 @@ const InputFlat = (props) => {
             value={props.value}
             name={props.fieldName}
           />
-          <small id={`${props.id}Help`} className="form-text text-muted">{props.supportText}</small>
+          <small id={`${props.id}Help`} className="form-text text-muted">
+            {props.supportText}
+          </small>
         </Column>
       </Row>
-
     </div>
-  )
-}
+  );
+};
 
 export default InputFlat;
