@@ -1,5 +1,5 @@
 import React from "react";
-import { firebase } from '../Firebase';
+import { firebase } from "../Firebase";
 
 const LoginButton = () => {
   return <button onClick={firebase.signInWithGoogle}>Log In</button>;
