@@ -1,15 +1,23 @@
-import React from 'react';
+import React from "react";
 
 const SelectField = (props) => {
-  let options = props.options.map(opt => {
-    return (<option key={opt} >{opt}</option>)
-  })
+  let options = props.options.map((opt) => {
+    return <option key={opt}>{opt}</option>;
+  });
 
-  if(props.flat) {
+  if (props.flat) {
     return (
-      <div className={`form-group ${props.center ? 'text-center' : ''}`}>
+      <div className={`form-group ${props.center ? "text-center" : ""}`}>
         <label htmlFor="exampleFormControlSelect1">{props.label}</label>
-        <select className={`form-control form-control-sm ${props.center ? 'text-center' : ''}`} id="exampleFormControlSelect1" value={props.value} onChange={props.onChange} name={props.fieldName}>
+        <select
+          className={`form-control form-control-sm ${
+            props.center ? "text-center" : ""
+          }`}
+          id="exampleFormControlSelect1"
+          value={props.value}
+          onChange={props.onChange}
+          name={props.fieldName}
+        >
           {options}
         </select>
       </div>
@@ -17,13 +25,21 @@ const SelectField = (props) => {
   }
 
   return (
-    <div className={`form-group ${props.center ? 'text-center' : ''}`}>
+    <div className={`form-group ${props.center ? "text-center" : ""}`}>
       <label htmlFor="exampleFormControlSelect1">{props.label}</label>
-      <select className={`form-control form-control-sm ${props.center ? 'text-center' : ''}`} id="exampleFormControlSelect1" value={props.value} onChange={props.onChange} name={props.fieldName}>
+      <select
+        className={`form-control form-control-sm ${
+          props.center ? "text-center" : ""
+        }`}
+        id="exampleFormControlSelect1"
+        value={props.value}
+        onChange={props.onChange}
+        name={props.fieldName}
+      >
         {options}
       </select>
     </div>
   );
-}
+};
 
 export default SelectField;
