@@ -1,5 +1,5 @@
 import React from "react";
-import { mount } from "enzyme";
+import { render } from "@testing-library/react";
 
 import Column from "../Column";
 
@@ -7,20 +7,16 @@ describe("Column", () => {
   const classProp = "test-class";
   const childContent = "Child Content";
 
-  const mountComp = overrides => {
-    return mount(
+  const mountComp = (overrides) => {
+    return render(
       <Column className={classProp} {...overrides}>
         {childContent}
       </Column>
     );
   };
 
-  let wrapper;
-
-  beforeEach(() => (wrapper = mountComp()));
-
-  it("renders a div with props.className in the className", () => {
-    expect(wrapper.text()).toContain(childContent);
-    expect(wrapper.props().className).toContain(classProp);
+  it("displays child content", () => {
+    const { getByText } = mountComp();
+    expect(getByText(childContent)).toBeInTheDocument();
   });
 });
