@@ -8,5 +8,5 @@ export { default as Row } from "./Row";
 export { default as SelectField } from "./SelectField";
 export { default as Textarea } from "./Textarea";
 export { default as LoginButton } from "./LoginButton";
-export { default as TopBar } from './TopBar';
+export { default as TopBar } from "./TopBar";
 export { default as Modal } from "./Modal";
