@@ -16,136 +16,165 @@ global.creatureTypes = [
 ];
 
 global.abilityScores = [
-  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30
-]
+  1,
+  2,
+  3,
+  4,
+  5,
+  6,
+  7,
+  8,
+  9,
+  10,
+  11,
+  12,
+  13,
+  14,
+  15,
+  16,
+  17,
+  18,
+  19,
+  20,
+  21,
+  22,
+  23,
+  24,
+  25,
+  26,
+  27,
+  28,
+  29,
+  30,
+];
 
 global.damageTypes = [
-  'Acid',
-  'Bludgeoning',
-  'Cold',
-  'Fire',
-  'Force',
-  'Lightning',
-  'Necrotic',
-  'Piercing',
-  'Poison',
-  'Psychic',
-  'Radiant',
-  'Slashing',
-  'Thunder'
-]
+  "Acid",
+  "Bludgeoning",
+  "Cold",
+  "Fire",
+  "Force",
+  "Lightning",
+  "Necrotic",
+  "Piercing",
+  "Poison",
+  "Psychic",
+  "Radiant",
+  "Slashing",
+  "Thunder",
+];
 
 global.skills = [
   {
-    name: 'Athletics',
-    ability: 'str'
+    name: "Athletics",
+    ability: "str",
   },
   {
-    name: 'Acrobatics',
-    ability: 'dex'
+    name: "Acrobatics",
+    ability: "dex",
   },
   {
-    name: 'Sleight of Hand',
-    ability: 'dex'
+    name: "Sleight of Hand",
+    ability: "dex",
   },
   {
-    name: 'Stealth',
-    ability: 'dex'
+    name: "Stealth",
+    ability: "dex",
   },
   {
-    name: 'Arcana',
-    ability: 'int'
+    name: "Arcana",
+    ability: "int",
   },
   {
-    name: 'History',
-    ability: 'int'
+    name: "History",
+    ability: "int",
   },
   {
-    name: 'Investigation',
-    ability: 'int'
+    name: "Investigation",
+    ability: "int",
   },
   {
-    name: 'Nature',
-    ability: 'int'
+    name: "Nature",
+    ability: "int",
   },
   {
-    name: 'Religion',
-    ability: 'int'
+    name: "Religion",
+    ability: "int",
   },
   {
-    name: 'Animal Handling',
-    ability: 'wis'
+    name: "Animal Handling",
+    ability: "wis",
   },
   {
-    name: 'Insight',
-    ability: 'wis'
+    name: "Insight",
+    ability: "wis",
   },
   {
-    name: 'Medicine',
-    ability: 'wis'
+    name: "Medicine",
+    ability: "wis",
   },
   {
-    name: 'Perception',
-    ability: 'wis'
+    name: "Perception",
+    ability: "wis",
   },
   {
-    name: 'Survival',
-    ability: 'wis'
+    name: "Survival",
+    ability: "wis",
   },
   {
-    name: 'Deception',
-    ability: 'cha'
+    name: "Deception",
+    ability: "cha",
   },
   {
-    name: 'Intimidation',
-    ability: 'cha'
+    name: "Intimidation",
+    ability: "cha",
   },
   {
-    name: 'Perfoemance',
-    ability: 'cha'
+    name: "Perfoemance",
+    ability: "cha",
   },
   {
-    name: 'Persuasion',
-    ability: 'cha'
+    name: "Persuasion",
+    ability: "cha",
   },
-]
+];
 
 global.languages = [
-  'Common',
-  'Dwarvish',
-  'Elvish',
-  'Giant',
-  'Gnomish',
-  'Goblin',
-  'Halfling',
-  'Orc',
-  'Abyssal',
-  'Celestial',
-  'Draconic',
-  'Deep Speech',
-  'Infernal',
-  'Aquan (Primordial)',
-  'Auran (Primordial)',
-  'Ignan (Primordial)',
-  'Terran (Primordial)',
-  'Sylvan',
-  'Undercommon'
-]
+  "Common",
+  "Dwarvish",
+  "Elvish",
+  "Giant",
+  "Gnomish",
+  "Goblin",
+  "Halfling",
+  "Orc",
+  "Abyssal",
+  "Celestial",
+  "Draconic",
+  "Deep Speech",
+  "Infernal",
+  "Aquan (Primordial)",
+  "Auran (Primordial)",
+  "Ignan (Primordial)",
+  "Terran (Primordial)",
+  "Sylvan",
+  "Undercommon",
+];
 
 global.conditions = [
-  'Blinded',
-  'Charmed',
-  'Deafened',
-  'Exhaustion',
-  'Fatigued',
-  'Frightened',
-  'Grappled',
-  'Incapacitated',
-  'Paralyzed',
-  'Petrified',
-  'Poisoned',
-  'Prone',
-  'Restrained',
-  'Stunned',
-  'Unconscious'
-]
+  "Blinded",
+  "Charmed",
+  "Deafened",
+  "Exhaustion",
+  "Fatigued",
+  "Frightened",
+  "Grappled",
+  "Incapacitated",
+  "Paralyzed",
+  "Petrified",
+  "Poisoned",
+  "Prone",
+  "Restrained",
+  "Stunned",
+  "Unconscious",
+];
